+++ Changelog.md
@@ -17,6 +17,7 @@ OpenCore Changelog
 - Fixed intermittent 32-bit prelinking failures caused by improper Mach-O expansion
 - Fixed failures in cacheless injection dependency resolution
 - Fixed detection issues with older Atom CPUs
+- Fixed `ScanPolicy` NVMe handling on MacPro5,1
 
 #### v0.6.2
 - Updated builtin firmware versions for SMBIOS and the rest